// src/types/core.ts
export type TimerMode = 'STOPWATCH' | 'TABATA'

/**
 * Defines the possible phases of the application timer.
 */
export type TimerPhase =
  | 'IDLE'
  | 'PREPARE'
  | 'WORK'
  | 'REST'
  | 'COOLDOWN'
  | 'RUNNING'

/**
 * Represents the complete state of the application timer.
 */
export interface TimerData {
  isRunning: boolean
  currentPhase: TimerPhase
  timeRemaining: number
  timeElapsed: number
  caloriesBurned: number
  mode: TimerMode
  workDuration: number
  restDuration: number
  soundToPlay?: 'WORK' | 'REST' | 'COUNTDOWN'
  soundEventId: number
}